+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Adapt fill-forms tests to DocumentServer v8.2
 * Simplify `DocumentServerVersion#supported_languages` method
 * Fix `rubocop-1.65.0` cop `Gemspec/AddRuntimeDependency`
+* Add closing circular references alert when open file 
 
 ### Tests